#582 commonization of compose and test modules for js
