+++ resources-compose/src/commonJsMain/kotlin/dev/icerock/moko/resources/compose/ImageResource.kt
