+++ resources-compose/src/commonJsMain/kotlin/dev/icerock/moko/resources/compose/internal/Platform.commonJs.kt
@@ -1,0 +1,7 @@
+/*
+ * Copyright 2025 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
+ */
+
+package dev.icerock.moko.resources.compose.internal
+
+internal expect suspend fun fetchByteArray(url: String): ByteArray
