+++ resources-test/src/commonJsMain/kotlin/dev/icerock/moko/resources/test/createImageResourceMock.kt
