--- samples/compose-resources-gallery/desktopApp/src/jsMain/kotlin/main.js.kt
@@ -1,12 +1,0 @@
-import androidx.compose.ui.ExperimentalComposeUiApi
-import androidx.compose.ui.window.CanvasBasedWindow
-import org.jetbrains.skiko.wasm.onWasmReady
-
-@OptIn(ExperimentalComposeUiApi::class)
-public fun main() {
-    onWasmReady {
-        CanvasBasedWindow("Js App") {
-            MainView()
-        }
-    }
-}
