--- samples/compose-resources-gallery/desktopApp/src/jsMain/resources/styles.css
@@ -1,8 +1,0 @@
-#root {
-    width: 100%;
-    height: 100vh;
-}
-
-#root > .compose-web-column > div {
-    position: relative;
-}
