--- samples/compose-resources-gallery/desktopApp/src/wasmJsMain/kotlin/main.js.kt
@@ -1,9 +1,0 @@
-import androidx.compose.ui.ExperimentalComposeUiApi
-import androidx.compose.ui.window.CanvasBasedWindow
-
-@OptIn(ExperimentalComposeUiApi::class)
-public fun main() {
-    CanvasBasedWindow("Js App") {
-        MainView()
-    }
-}
