--- samples/compose-resources-gallery/desktopApp/src/wasmJsMain/resources/index.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Kamel Samples</title>
-    <script src="skiko.js"> </script>
-    <link type="text/css" rel="stylesheet" href="styles.css" />
-</head>
-<body>
-<div>
-    <canvas id="ComposeTarget" width="600" height="600"></canvas>
-</div>
-<script src="desktopApp.js"> </script>
-</body>
-</html>
